Split the semihosting options from the semihosting behaviour
