+++ probe-rs-tools/src/bin/probe-rs/rpc/functions.rs
@@ -101,6 +101,7 @@ pub mod reset;
 pub mod resume;
 pub mod rtt_client;
 pub mod rtt_config;
+pub mod semihosting_options;
 pub mod stack_trace;
 pub mod test;
