+++ probe-rs-tools/src/bin/probe-rs/rpc/utils/semihosting.rs
@@ -1,4 +1,5 @@
 use crate::rpc::functions::monitor::SemihostingEvent;
+pub use crate::rpc::functions::semihosting_options::{Mapping, SemihostingOptions};
 use probe_rs::{
     Core,
     semihosting::{
@@ -24,18 +25,7 @@ enum FileHandle {
     UnixStream(UnixStream),
 }
 
-use std::convert::Infallible;
-
-use postcard_schema::Schema;
 use regex::Regex;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Schema, Clone)]
-enum Mapping {
-    Exact(String, String),
-    Prefix(String, String),
-    Regex(String, String),
-}
 
 enum FileVariant {
     File(String),
@@ -43,76 +33,49 @@ enum FileVariant {
     UnixStream(String),
 }
 
-#[derive(Serialize, Deserialize, Schema, Clone)]
-pub struct SemihostingOptions {
-    mappings: Vec<Mapping>,
-}
-
-impl SemihostingOptions {
-    pub fn new() -> Self {
-        Self { mappings: vec![] }
-    }
-
-    pub fn add_file(&mut self, from: String, to: String) -> Result<(), Infallible> {
-        self.mappings.push(Mapping::Exact(from, to));
-        Ok(())
-    }
-
-    pub fn add_file_prefix(&mut self, from: String, to: String) -> Result<(), Infallible> {
-        self.mappings.push(Mapping::Prefix(from, to));
-        Ok(())
-    }
-
-    pub fn add_file_regex(&mut self, re: String, to: String) -> Result<(), regex::Error> {
-        let _ = Regex::new(&re)?; // Check if it's a valid regular expression
-        self.mappings.push(Mapping::Regex(re, to));
-        Ok(())
-    }
-
-    fn map_file(&self, value: &str) -> Option<String> {
-        for item in &self.mappings {
-            match item {
-                Mapping::Exact(from, to) => {
-                    if value == from {
-                        return Some(to.clone());
-                    }
+fn map_file(options: &SemihostingOptions, value: &str) -> Option<String> {
+    for item in options.mappings() {
+        match item {
+            Mapping::Exact(from, to) => {
+                if value == from {
+                    return Some(to.clone());
                 }
-                Mapping::Prefix(prefix, to) => {
-                    if let Some(rest) = value.strip_prefix(prefix) {
-                        return Some(to.clone() + rest);
-                    }
+            }
+            Mapping::Prefix(prefix, to) => {
+                if let Some(rest) = value.strip_prefix(prefix) {
+                    return Some(to.clone() + rest);
                 }
-                Mapping::Regex(re, to) => {
-                    let re = Regex::new(re).expect("valid Regex");
-                    if let Some(captures) = re.captures(value) {
-                        let mut ret = String::new();
-                        captures.expand(to, &mut ret);
-                        return Some(ret);
-                    }
+            }
+            Mapping::Regex(re, to) => {
+                let re = Regex::new(re).expect("valid Regex");
+                if let Some(captures) = re.captures(value) {
+                    let mut ret = String::new();
+                    captures.expand(to, &mut ret);
+                    return Some(ret);
                 }
             }
         }
-
-        None
     }
 
-    fn find_file(&self, value: &str) -> Option<FileVariant> {
-        let v = self.map_file(value)?;
+    None
+}
 
-        if let Some(r) = v.strip_prefix("file:") {
-            return Some(FileVariant::File(r.into()));
-        }
+fn find_file(options: &SemihostingOptions, value: &str) -> Option<FileVariant> {
+    let v = map_file(options, value)?;
 
-        if let Some(r) = v.strip_prefix("tcp:") {
-            return Some(FileVariant::TcpStream(r.into()));
-        }
-
-        if let Some(r) = v.strip_prefix("unix:") {
-            return Some(FileVariant::UnixStream(r.into()));
-        }
-
-        Some(FileVariant::File(v))
+    if let Some(r) = v.strip_prefix("file:") {
+        return Some(FileVariant::File(r.into()));
     }
+
+    if let Some(r) = v.strip_prefix("tcp:") {
+        return Some(FileVariant::TcpStream(r.into()));
+    }
+
+    if let Some(r) = v.strip_prefix("unix:") {
+        return Some(FileVariant::UnixStream(r.into()));
+    }
+
+    Some(FileVariant::File(v))
 }
 
 pub struct SemihostingFileManager {
@@ -260,7 +223,7 @@ impl SemihostingFileManager {
 
         let f = if path == ":tt" {
             self.open_tt(request.mode())
-        } else if let Some(path) = self.semihosting_options.find_file(&path) {
+        } else if let Some(path) = find_file(&self.semihosting_options, &path) {
             match path {
                 FileVariant::File(path) => self.open_file(&path, request.mode()),
                 FileVariant::TcpStream(addr) => self.open_tcp_stream(&addr),
